backend/requirements.txt
sentence-transformers==2.2.2
chromadb==0.4.22
langchain==0.1.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0